[feat] 회원가입 완료 후 알림 토큰 발급 (#243)
