+++ Sukbakji/PresentationLayer/Sources/Home/View/Home/HomeViewController.swift
@@ -18,7 +18,6 @@ class HomeViewController: UIViewController, View {
     private let favoriteBoardViewModel = FavoriteBoardViewModel()
     private let hotPostViewModel = HotPostViewModel()
     private let favoriteLabViewModel = FavoriteLabViewModel()
-    private let myProfileViewModel = MyProfileViewModel()
     var disposeBag = DisposeBag()
     var reactor: HomeReactor?
     
@@ -33,7 +32,6 @@ class HomeViewController: UIViewController, View {
         setAPI()
         self.reactor = HomeReactor()
         bind(reactor: reactor!)
-        getFCMToken()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -51,19 +49,6 @@ extension HomeViewController {
         homeView.mypageButton.addTarget(self, action: #selector(info_Tapped), for: .touchUpInside)
         homeView.adCollectionView.delegate = self
         homeView.adCollectionView.dataSource = self
-    }
-    
-    private func getFCMToken() {
-        Messaging.messaging().token { token, error in
-            if let error = error {
-                print("FCM 토큰 가져오기 실패: \(error.localizedDescription)")
-            } else if let token = token {
-                print("현재 FCM 토큰: \(token)")
-                // 서버에 토큰 업로드
-                TokenManager.shared.saveFCMToken(token)
-                self.myProfileViewModel.uploadFCMTokenToServer(fcmToken: token)
-            }
-        }
     }
     
     private func setAPI() {
